// src/shared/api/models/getApiStatisticsAppointments200ItemsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsAppointments200ItemsItem = {
    id: number;
    customerName: string;
    serviceName: string;
    dateTime: string;
    status: string;
    totalAmount: number;
};