// src/shared/api/models/getApiStatisticsServices200ServicesItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsServices200ServicesItem = {
    id: number;
    serviceId: number;
    serviceName: string;
    numberOfBookings: number;
    revenue: number;
    percentage: number;
};