// src/shared/api/models/getApiStatisticsServicesStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiStatisticsServicesStatus =
    (typeof GetApiStatisticsServicesStatus)[keyof typeof GetApiStatisticsServicesStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiStatisticsServicesStatus = {
    ALL: 'ALL',
    BOOKED: 'BOOKED',
    RESCHEDULED: 'RESCHEDULED',
    COMPLETED: 'COMPLETED',
    CANCELLED: 'CANCELLED',
    OFFERED: 'OFFERED',
    OFFER_ACCEPTED: 'OFFER_ACCEPTED',
    DELETED: 'DELETED',
    NOSHOW: 'NOSHOW',
    AUTOCOMPLETED: 'AUTOCOMPLETED',
} as const;